::major overhaul of how the code is written:: there are 2 major functions now:       readhead - simple data return for header values       readsnap - reads snapshot data depending on arguments provided
adjusted the code to be more modular.  Now each block skip is a module which makes it an easier read.  Also added a conversion switch for RHO & U.  If supplied units=1 then it will return RHO in cgs(g/cm^3), and U in kelvin.

**CURRENTLY TEMPERATURE RETURN IS NOT FUNCTIONING PROPERLY**

BUGS: right now when pointing to a multi-file snapshot without specifying numfiles the code seg faults.